[WIP] implementation of Optional types for #66
manifold-lang/manifold-core/issues/35 was preventing code from
deserializing an `Array(Bool)` or `Optional(Bool)`,
because `BooleanTypeValue` and `UserDefinedTypeValue(BooleanTypeValue)` are different types.

There's probably still bugs related to `UserDefinedTypeValue`s in this
code (the optional values aren't the same type as the claimed type).

We will probably still need to be able to deserialize `Bit`s, where `type Bit = Bool`.

-	Can/Should the compiler dereference all of the aliases in a user defined
	type value when deserializing?
-	If not, is there a need `UserDefinedTypeValue` as well, to distinguish
	between Bit and Bool
	(Are UDTs more like haskell types or like C `typedef`s.
	I can't tell from the documentation.)

Also, it seems like AttributeTypes have to be a string (UDT name or
primitive) when they are being deserialized in `SchematicDeserializer`.
Should this be changed to allow `{type:"Array"/"Optional", elementType: ...}`